+++ src/eca/server.clj
@@ -94,6 +94,9 @@
 (defmethod jsonrpc.server/receive-notification "chat/promptStop" [_ components params]
   (handlers/chat-prompt-stop (with-config components) params))
 
+(defmethod jsonrpc.server/receive-notification "chat/promptSteer" [_ components params]
+  (handlers/chat-prompt-steer (with-config components) params))
+
 (defmethod jsonrpc.server/receive-request "chat/delete" [_ components params]
   (eventually (handlers/chat-delete (with-config components) params)))